adding 'manage' option to Configure::DSL
This will allow you to add the following line to your .dotrc file:

    manage '/some/random/path/to/file.ext'

This will setup a file in Dotify called 'file.ext' and which is the
source of the now-symlinked '/some/random/path/to/file.ext'.

If, for some reason, you have two different paths that shared the same
filename and extension, this would could both destination files to be
symlinked to the same file. This may be desired, but often there are
two different files entirely. You would do it this way:

    manage '/some/random/path/to/file.ext', 'example/file.ext'

The second parameter is the path within Dotify to the file soruce.
When Dotify syncs all of the files, it will create the
`~/.dotify/example` directory and then created the `file.ext` within
that directory.