prevent adding duplicate Pointer to FileList
This prevents any new pointers that are manually
defined by the user in the .dotrc configuration from
overwriting previously linked file references.

In essence, if FileList attempts to add a new Pointer,
but the Pointer source has already been defined on the list,
it should let the first Pointer source win out and the second
Pointer is ignored.